requirements.txt
customtkinter
Pillow
pyautogui        
interception-python
easyocr
torch
numpy              
opencv-python